+++ lib/circleci/coverage_reporter.rb
@@ -3,10 +3,10 @@ require_relative 'coverage_reporter/configuration'
 require_relative 'coverage_reporter/errors'
 require_relative 'coverage_reporter/runner'
 
-require_relative 'coverage_reporter/flow/reporter'
-require_relative 'coverage_reporter/link/reporter'
-require_relative 'coverage_reporter/rubycritic/reporter'
-require_relative 'coverage_reporter/simplecov/reporter'
+require_relative 'coverage_reporter/reporters/flow'
+require_relative 'coverage_reporter/reporters/link'
+require_relative 'coverage_reporter/reporters/rubycritic'
+require_relative 'coverage_reporter/reporters/simplecov'
 
 module CircleCI
   module CoverageReporter