+++ lib/circleci/coverage_reporter/report.rb
@@ -1,22 +1,13 @@
 module CircleCI
   module CoverageReporter
+    # Encapsulate a report created by a reporter.
+    #
+    # @see Reporters::Base#report
     class Report
-      # @return [AbstractReporter]
-      attr_reader :reporter
-
-      # @return [AbstractResult]
-      attr_reader :current_result
-
-      # @return [AbstractResult, nil]
-      attr_reader :base_result
-
-      # @return [AbstractResult, nil]
-      attr_reader :previous_result
-
-      # @param reporter [AbstractReporter]
-      # @param current [AbstractCurrentResult]
-      # @param base [AbstractBuildResult, nil]
-      # @param previous [AbstractBuildResult, nil]
+      # @param reporter [Reporters::Base] the reporter of the report
+      # @param current [Result]
+      # @param base [Result, nil] result at master branch
+      # @param previous [Result, nil] result at previous build in same branch
       def initialize(reporter, current, base: nil, previous: nil)
         @reporter = reporter
         @current_result = current
@@ -24,24 +15,49 @@ module CircleCI
         @previous_result = previous
       end
 
-      # @return [String] coverage reporter name
-      def type
-        reporter.name
+      # @return [String]
+      def to_s
+        "#{link}: #{current_result.pretty_coverage}#{emoji}#{progress}"
+      end
+
+      private
+
+      attr_reader :reporter, :current_result, :base_result, :previous_result
+
+      # @return [String]
+      def link
+        "[#{reporter.name}](#{current_result.url})"
       end
 
       # @return [String]
-      def pretty_coverage
-        current_result.pretty_coverage
+      def emoji
+        if base_diff.nil? || base_diff.nan? || base_diff.round(2).zero?
+          ''
+        elsif base_diff.positive?
+          ':chart_with_upwards_trend:'
+        else
+          ':chart_with_downwards_trend:'
+        end
       end
 
-      # @return [Float] coverage percent of the current build result
-      def coverage
-        current_result.coverage
+      # @return [String]
+      def progress
+        progresses.empty? ? '' : "(#{progresses.join(', ')})"
       end
 
-      # @return [String] URL for current coverage build result
-      def url
-        current_result.url
+      # @return [Array<String>]
+      def progresses
+        [base_progress, branch_progress].compact
+      end
+
+      # @return [String, nil]
+      def base_progress
+        base_diff ? "[master](#{base_result.url}): #{pretty_base_diff}" : nil
+      end
+
+      # @return [String, nil]
+      def branch_progress
+        branch_diff ? "[previous](#{previous_result.url}): #{pretty_branch_diff}" : nil
       end
 
       # @return [String, nil]
@@ -67,8 +83,6 @@ module CircleCI
         return unless previous_result
         current_result.coverage - previous_result.coverage
       end
-
-      private
 
       # @param diff [Float]
       # @return [String]