+++ spec/circleci/coverage_reporter/client_spec.rb
@@ -1,4 +1,0 @@
-require 'circleci/coverage_reporter/client'
-
-RSpec.describe CircleCI::CoverageReporter::Client do
-end