+++ spec/circleci/coverage_reporter/configuration_spec.rb
@@ -1,4 +1,0 @@
-require 'circleci/coverage_reporter/configuration'
-
-RSpec.describe CircleCI::CoverageReporter::Configuration do
-end